pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.74.0",
    "fastapi>=0.121.2",
    "firecrawl-py>=1.5.0",
    "langchain>=1.0.7",
    "langchain-google-genai>=3.0.3",
    "langgraph>=1.0.3",
    "openai>=2.8.1",
    "pydantic-settings>=2.12.0",
    "uvicorn[standard]>=0.38.0",
]
